Create a new common class 'BaseGlobalState' for 'MaterializeGlobalState' and 'PostgresGlobalState' to reduce code duplication
